requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
google-generativeai>=0.3.0
python-decouple>=3.5
django-cors-headers